# app/swagger/livrable.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "EndPoints pour livrable",
        "version": "1.0.0"
    },
    "paths": {
        "/api/livrable": {
            "get": {
                "summary": "Retourne tout les livrables",
                "operationId": "f142d42d407a1844efdaccd8c5c79bce",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "api/livrable/create": {
            "post": {
                "summary": "Ajout d'un livrable",
                "operationId": "5dba6eda9ee2668a230ec6a6fcbf57e2",
                "responses": {
                    "201": {
                        "description": "Livrable créé avec succes"
                    },
                    "422": {
                        "description": "livrable introuvable"
                    }
                }
            }
        },
        "/api/livrable{id}": {
            "get": {
                "summary": "Afficher un livrable",
                "operationId": "6fbfbb3b22e34b614e5af9f44174e240",
                "responses": {
                    "200": {
                        "description": "succes"
                    },
                    "400": {
                        "description": "erreur"
                    }
                }
            }
        },
        "/ap/livrable/{id}": {
            "patch": {
                "summary": "Modification d'un guide",
                "operationId": "b5d705d20ad3c099d88c9b8121245f24",
                "responses": {
                    "201": {
                        "description": "Livrable modifié avec succes"
                    },
                    "422": {
                        "description": "Erreur"
                    }
                }
            }
        },
        "/api/livrable/{id}": {
            "delete": {
                "summary": "Suppression d'un livrable",
                "operationId": "ed73de3d2c643ff46d25d41018c9f755",
                "responses": {
                    "200": {
                        "description": "Livrable supprimé avec succes"
                    },
                    "422": {
                        "description": "erreur"
                    }
                }
            }
        }
    }
}